tests/wpt/web-platform-tests/largest-contentful-paint/cross-origin-image.sub.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: observe cross-origin images but without renderTime.</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<script>
  async_test(function (t) {
    if (!window.LargestContentfulPaint) {
      assert_unreached("LargestContentfulPaint is not implemented");
    }
    const beforeLoad = performance.now();
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        const url = 'http://{{domains[www]}}:{{ports[http][1]}}/images/blue.png';
        // blue.png is 133 x 106.
        const size = 133 * 106;
        checkImage(entry, url, 'image_id', size, beforeLoad, ['renderTimeIs0']);
      })
    );
    observer.observe({type: 'largest-contentful-paint', buffered: true});
  }, 'Cross-origin image is observable, with renderTime equal to 0.');
</script>

<img src='http://{{domains[www]}}:{{ports[http][1]}}/images/blue.png' id='image_id'/>
</body>
